How to make subclass factory?)
/
