+++ src/Objects.py
@@ -6,6 +6,7 @@ from registry import DefaultBlockRegistry as registry
 from loguru import logger
 
 
+
 dirt_model = Model('dirt', color(0,0,0), 'cube')
 cobblestone_model = Model('cobblestone1', color(0,0,0), 'cube')
 
@@ -17,7 +18,7 @@ class dirt(Block):
         self.texture = dirt_model.texture
         self.color = dirt_model.color
         self.model = dirt_model.model
-        self.item = Item(self.name, self.id, self.texture, dirt_model)
+        self.item = itemfactory(self.name, self.id, self.texture, dirt_model)
         self.item.isBlockItem = True
         self.item.block = self
 
@@ -29,7 +30,7 @@ class cobblestone(Block):
         self.texture = cobblestone_model.texture
         self.color = cobblestone_model.color
         self.model = cobblestone_model.model
-        self.item = Item(self.name, self.id, self.texture, dirt_model)
+        self.item = itemfactory(self.name, self.id, self.texture, dirt_model)
         self.item.isBlockItem = True
         self.item.block = self
