+++ src/VoxelTypes.py
@@ -65,118 +65,23 @@ class Crosshair(Entity):
         )
         self.enabled = True
 
-class PauseMenu(Entity):
-    def __init__(self, player, wrld):
-        logger.info("Pause menu has been initalized")
-        self.player = player
-        super().__init__(
-            parent=camera.ui,
-            enabled=False,
-            model='quad',
-            texture='white_cube',
-            color=color(0, 0, 0, 0.8),
-            scale=(1, 2),
-            position=(0, 0, -1)
-        )
 
-        self.exit_button = Button(
-            parent=self,
-            position=(0, -0.05),
-            scale=(0.5, 0.05),
-            text='Exit',
-            on_click=self.exit_game
-        )
-
-        self.close_button = Button(
-            parent=self,
-            position=(0, 0.05),
-            scale=(0.5, 0.05),
-            text='Continue',
-            on_click=self.close_menu
-        )
-
-        #self.save_textbox = TextField(
-        #    parent=self,
-        #    position=(0, -0.1),
-        #    scale=(0.1, 1),
-        #    line_height=0.5,
-        #    max_lines = 1,
-        #    color = None
-        #)
-
-    
-        self.wrld = wrld
-    def exit_game(self):
-        exit()
-
-    def close_menu(self):
-        self.enabled = False
-        mouse.locked = True
-        self.player.enabled = True
-    def save_game(self):
-        self.wrld.Save(self.save_textbox.text)
-    def load_game(self):
-        self.wrld.Unload()
-        self.wrld.Load(self.save_textbox.text)
     
 
 class Item():
-    def __init__(self, name, id, invtext, model, description=""):
-        self.name = name
-        self.id = id
-        self.invtext = invtext
-        self.model = model
-        self.isBlockItem = False
-        self.whenClicked = lambda: logger.debug(f"Item {self.name} has been interacted with")
+    #def __init__(self, name, id, invtext, model, description=""):
+    #    self.name = name
+    #    self.id = id
+    #    self.invtext = invtext
+    #    self.model = model
+    #    self.isBlockItem = False
+    #    self.whenClicked = lambda: logger.debug(f"Item {self.name} has been interacted with")
     def __init_subclass__(self):
         logger.debug(f"Item {self.id} has been defined")
 
 
 
-class Hotbar(Entity):
-    def __init__(self, num_slots=10):
-        super().__init__(parent=camera.ui, scale=0.1, position=(-0.45, -0.4))
-        self.num_slots = num_slots
-        self.selected_slot = 0
-        self.count = [0,0,0,0,0,0,0,0,0,0]
-        self.items = [Item("None", "null", None, None)] * num_slots  # Initialize with default items
 
-        # Create slots using item.invtext for the texture
-        self.slots = []
-        for i in range(num_slots):
-            slot = Button(
-                parent=self,
-                model='quad',
-                texture=self.items[i].invtext,
-                color=color(1, 1, 1),
-                position=(0.8 * i, 0),
-                scale=0.8,  # Adjust the scale as needed for your textures
-                on_click=Func(self.select_slot, i)
-            )
-            self.slots.append(slot)
-
-        # Highlight the selected slot
-        self.highlight_selected_slot()
-
-    def select_slot(self, slot_index):
-        self.selected_slot = slot_index
-        self.highlight_selected_slot()
-
-    def highlight_selected_slot(self):
-        for i, slot in enumerate(self.slots):
-            if i == self.selected_slot:
-                slot.color = white  # Highlighted color
-            else:
-                slot.color = gray  # Default color
-
-    def add_item(self, item, count, slot_index):
-        self.items[slot_index] = item
-        if item.isBlockItem == True:
-            self.slots[slot_index].texture = item.invtext
-        else:
-            self.slots[slot_index].texture = item.invtext
-        self.slots[slot_index].texture = item.invtext  # Update the texture
-        self.count[slot_index] = count
     
 class ItemRegistry():
     items = {}
